refactor: move zstd module to file
